Change type of tile_size in dwconv ukernel
uint8_t is too small for tile size, since for multipass this can be very large if the middle pass is run multiple times (e.g. large kernel size of 513x1 seen in the wild).

PiperOrigin-RevId: 565542870

+++ src/xnnpack/operator.h
@@ -44,7 +44,7 @@ struct xnn_ukernel_dwconv {
   uint8_t last_tile;
   // For unipass, tile_size == primary_tile, otherwise it is calculated based on
   // how many pass the middle_tile runs.
-  uint8_t tile_size;
+  size_t tile_size;
 };
 
 // Direct 2D Depthwise Convolution
